Reorganize CellTypeUI component files
