+++ frontend/src/Project/EditControls/CellTypeControls/CellTypeUI/CellTypeAccordion/CellGrid.js
@@ -4,8 +4,8 @@ import FormLabel from '@mui/material/FormLabel';
 import { useSelector } from '@xstate/react';
 import equal from 'fast-deep-equal';
 import { useState } from 'react';
-import { useLayers, useRaw } from '../../../ProjectContext';
-import { hexToRgb } from '../../../service/labels/cellTypesMachine';
+import { useLayers, useRaw } from '../../../../ProjectContext';
+import { hexToRgb } from '../../../../service/labels/cellTypesMachine';
 import AddCellsButton from './AddCellsButton';
 import ChannelChip from './ChannelChip';
 import ColoredChannelChip from './ColoredChannelChip';
